base model of prereg planner initialized
Additions
- able to generate table from the course selection menu
- able to export table as image

Changes
- table rendering only depends on QPainter
- TableWidget rows relies on updateHiddenLoadedCourses method

Known bugs
- course combination length of zero leads to program crashing

+++ Prereg_Engines/html_parsers.py
@@ -46,7 +46,7 @@ class HTMLCPParser:
             if not code in dict_course:
                 course.code = code
                 course.title = datas[2].text
-                course.ch = int(datas[3].text)
+                course.ch = float(datas[3].text)
                 dict_course[code] = course
 
             sect = CourseSection()
